requirements.txt
tensorflow>=2.10
numpy
opencv-python
pillow
matplotlib
scikit-learn
streamlit
tqdm
